SDI-2223–308-session10-Ejercicio complementario 1.

+++ musicstoreapp/routes/api/songsAPIv1.0.js
@@ -40,17 +40,32 @@ module.exports = function (app, songsRepository, usersRepository) {
         try {
             let songId = ObjectId(req.params.id)
             let filter = {_id: songId}
-            songsRepository.deleteSong(filter, {}).then(result => {
-                if (result === null || result.deletedCount === 0) {
-                    res.status(404);
-                    res.json({error: "ID inválido o no existe, no se ha borrado el registro."});
+
+            isAuthor(res.user, songId, function(author) {
+                if (author) {
+                    songsRepository.deleteSong(filter, {}).then(result => {
+                        if (result === null || result.deletedCount === 0) {
+                            res.status(404);
+                            res.json({error: "ID inválido o no existe, no se ha borrado el registro."});
+                        } else {
+                            res.status(200);
+                            res.send(JSON.stringify(result));
+                        }
+                    }).catch(error => {
+                        res.status(500);
+                        res.json({error: "Se ha producido un error al eliminar la canción."})
+                    });
                 } else {
-                    res.status(200);
-                    res.send(JSON.stringify(result));
+                    let errors = new Array();
+                    errors.push({
+                        "value": res.user,
+                        "msg": "Sólo el autor de la canción puede borrarla.",
+                        "param": "user",
+                        "location": "body"
+                    });
+                    res.status(422);
+                    res.json({errors: errors});
                 }
-            }).catch(error => {
-                res.status(500);
-                res.json({error: "Se ha producido un error al eliminar la canción."})
             });
         } catch (e) {
             res.status(500);
@@ -64,19 +79,27 @@ module.exports = function (app, songsRepository, usersRepository) {
                 title: req.body.title,
                 kind: req.body.kind,
                 price: req.body.price,
-                author: req.session.user
+                author: res.user
             }
-            // Validar aquí: título, género, precio y autor.
-            songsRepository.insertSong(song, function (songId) {
-                if (songId === null) {
-                    res.status(409);
-                    res.json({error: "No se ha podido crear la canción. El recurso ya existe."});
+            validatorInsertSong(song, function (errors) {
+                if (errors !== null && errors.length > 0) {
+                    res.status(422);
+                    res.json({errors: errors});
                 } else {
-                    res.status(201);
-                    res.json({
-                        message: "Canción añadida correctamente.",
-                        _id: songId
-                    })
+                    songsRepository.insertSong(song, function (songId) {
+                        if (songId === null) {
+                            res.status(409);
+                            let errors = new Array();
+                            errors.push("No se ha podido crear la canción. El recurso ya existe.");
+                            res.json({errors: errors});
+                        } else {
+                            res.status(201);
+                            res.json({
+                                message: "Canción añadida correctamente.",
+                                _id: songId
+                            })
+                        }
+                    });
                 }
             });
         } catch (e) {
@@ -89,7 +112,7 @@ module.exports = function (app, songsRepository, usersRepository) {
         try {
             let songId = ObjectId(req.params.id);
             let filter = {_id: songId};
-//Si la _id NO no existe, no crea un nuevo documento.
+            //Si la _id NO no existe, no crea un nuevo documento.
             const options = {upsert: false};
             let song = {
                 author: req.session.user
@@ -100,26 +123,45 @@ module.exports = function (app, songsRepository, usersRepository) {
                 song.kind = req.body.kind;
             if (typeof req.body.price !== "undefined" && req.body.price !== null)
                 song.price = req.body.price;
-            songsRepository.updateSong(song, filter, options).then(result => {
-                if (result === null) {
-                    res.status(404);
-                    res.json({error: "ID inválido o no existe, no se ha actualizado la canción."});
-                }
-                //La _id No existe o los datos enviados no difieren de los ya almacenados.
-                else if (result.modifiedCount == 0) {
-                    res.status(409);
-                    res.json({error: "No se ha modificado ninguna canción."});
-                }
-                else{
-                    res.status(200);
-                    res.json({
-                        message: "Canción modificada correctamente.",
-                        result: result
+
+            isAuthor(song.author, songId, function(author) {
+                if (author) {
+                    validatorInsertSong(song, function(errors) {
+                        if (errors == null || errors.length <= 0) {
+                            songsRepository.updateSong(song, filter, options).then(result => {
+                                if (result === null) {
+                                    res.status(404);
+                                    res.json({error: "ID inválido o no existe, no se ha actualizado la canción."});
+                                }
+                                //La _id No existe o los datos enviados no difieren de los ya almacenados.
+                                else if (result.modifiedCount == 0) {
+                                    res.status(409);
+                                    res.json({error: "No se ha modificado ninguna canción."});
+                                }
+                                else{
+                                    res.status(200);
+                                    res.json({
+                                        message: "Canción modificada correctamente.",
+                                        result: result
+                                    })
+                                }
+                            }).catch(error => {
+                                res.status(500);
+                                res.json({error : "Se ha producido un error al modificar la canción."})
+                            });
+                        }
                     })
+                } else {
+                    let errors = new Array();
+                    errors.push({
+                        "value": song.author,
+                        "msg": "Sólo el autor de la canción puede modificar sus datos",
+                        "param": "user",
+                        "location": "body"
+                    });
+                    res.status(403);
+                    res.json({errors: errors});
                 }
-            }).catch(error => {
-                res.status(500);
-                res.json({error : "Se ha producido un error al modificar la canción."})
             });
         } catch (e) {
             res.status(500);
@@ -170,4 +212,52 @@ module.exports = function (app, songsRepository, usersRepository) {
             });
         }
     });
+
+    function validatorInsertSong(song, callbackFunction) {
+        let errors = new Array();
+
+        if (song.title === null || typeof song.title === 'undefined' || song.title.trim().length === 0) {
+            errors.push({
+                "value": song.title,
+                "msg": "El título de la canción no puede estar vacío",
+                "param": "title",
+                "location": "body"
+            });
+        }
+        if (song.kind === null || typeof song.kind === 'undefined' || song.kind.trim().length === 0) {
+            errors.push({
+                "value": song.kind,
+                "msg": "El género de la canción no puede estar vacío",
+                "param": "kind",
+                "location": "body"
+            });
+        }
+        if (song.price === null || typeof song.price === 'undefined' || song.price <= 0
+                || song.price.trim().length === 0) {
+            errors.push({
+                "value": song.price,
+                "msg": "El precio de la canción debe ser mayor o igual que cero",
+                "param": "price",
+                "location": "body"
+            });
+        }
+
+        if (errors.length <= 0) {
+            callbackFunction(null);
+        } else {
+            callbackFunction(errors);
+        }
+    }
+
+    function isAuthor(user, songId, callbackFunction) {
+        let filterAuthor = {$and: [{"_id": songId}, {"author": user}]};
+        let options = {};
+        songsRepository.getSongs(filterAuthor, options).then(songs => {
+            if (songs == null || songs.length <= 0) {
+                callbackFunction(false);
+            } else {
+                callbackFunction(true);
+            }
+        });
+    }
 }
